Update to configuration options plus minor fixes
Allows to specify and/or override defaults for doorbell press, motion detected and person detected on non-HKSV cooldown times. The following are the new options supported in the config JSON:

    "MotionCooldown"
    "PersonCooldown"
    "DoorbellCooldown"

Values accepted are a number. If value is less than 1000, the value is treated as a seconds value rather than milliseconds